moved command.py to lib folder
